Remove all props and state from Activity.jsx
